refactor(spi): move SPI logic out of Memory.hs
- move SPI memory access handling and wave diagrams render into `Wrench.Machine.Spi`
- model SPI input values as frames with explicit bit length
- make `byte` use 8 SPI ticks and `word` use 32 SPI ticks
- expand `bytes` as consecutive 8-bit frames
- reject overlapping SPI input frames in config
