+++ src/wrench/Wrench/Machine/Memory.hs
@@ -11,11 +11,11 @@ module Wrench.Machine.Memory (
     prettyDump,
 ) where
 
-import Data.Bits (Bits (..), FiniteBits (..))
 import Numeric (showHex)
 import Relude
 import Relude.Extra
 import Relude.Unsafe qualified as Unsafe
+import Wrench.Machine.Spi
 import Wrench.Machine.Types
 import Wrench.Translator.Types
 
@@ -195,247 +195,6 @@ ioPortByteCollision io addr =
 ioWordKeys :: forall w isa. IoMem isa w -> [Int]
 ioWordKeys IoMem{mIoKeys, mSpiDevices} =
     mIoKeys <> spiWordKeys mSpiDevices
-
-spiWordKeys :: IntMap (SpiDevice w) -> [Int]
-spiWordKeys devices = ordNub $ concatMap spiDeviceAddrs (elems devices)
-
-spiDeviceAddrs :: SpiDevice w -> [Int]
-spiDeviceAddrs SpiDevice{spiPins = SpiPinsConf{spCsAddr, spClkAddr, spMosiAddr, spMisoAddr}} =
-    [spCsAddr, spClkAddr, spMosiAddr, spMisoAddr]
-
-spiOutputAddrs :: IntMap (SpiDevice w) -> [Int]
-spiOutputAddrs devices = ordNub $ concatMap spiDeviceOutputAddrs (elems devices)
-
-spiDeviceOutputAddrs :: SpiDevice w -> [Int]
-spiDeviceOutputAddrs SpiDevice{spiPins = SpiPinsConf{spCsAddr, spClkAddr, spMosiAddr}} =
-    [spCsAddr, spClkAddr, spMosiAddr]
-
-findSpiByteRegister :: forall w isa. (ByteSizeT w) => IoMem isa w -> Int -> Maybe (Int, Int)
-findSpiByteRegister IoMem{mSpiDevices} addr =
-    asum $ map match (spiWordKeys mSpiDevices)
-    where
-        wordSize = byteSizeT @w
-        match wordAddr
-            | wordAddr <= addr && addr < wordAddr + wordSize =
-                Just (addr - wordAddr, wordAddr)
-            | otherwise =
-                Nothing
-
-readSpiWord :: forall w isa. (MachineWord w) => IoMem isa w -> Int -> Maybe (Either Text (IoMem isa w, w))
-readSpiWord io@IoMem{mSpiDevices} addr =
-    if addr `elem` spiWordKeys mSpiDevices
-        then Just $ Right (io, spiPinsWord addr mSpiDevices)
-        else Nothing
-
-writeSpiWord :: forall w isa. (MachineWord w) => IoMem isa w -> Int -> w -> Maybe (Either Text (IoMem isa w))
-writeSpiWord io@IoMem{mSpiDevices, mClock} addr word
-    | addr `notElem` spiWordKeys mSpiDevices = Nothing
-    | addr `notElem` spiOutputAddrs mSpiDevices =
-        Just $ Left $ "iomemory[" <> show addr <> "]: can't write to SPI input-only pin address"
-    | otherwise =
-        let devices' = fmap (writeSpiPins mClock addr word) mSpiDevices
-         in Just $ Right io{mSpiDevices = devices'}
-
-readSpiByte :: forall w isa. (MachineWord w) => IoMem isa w -> Int -> Maybe (Either Text (IoMem isa w, Word8))
-readSpiByte io addr =
-    findSpiByteRegister @w io addr <&> \(offset, wordAddr) -> do
-        case readSpiWord io wordAddr of
-            Just result -> do
-                (io', word) <- result
-                return (io', wordSplit word Unsafe.!! offset)
-            Nothing -> Left $ "iomemory[" <> show addr <> "]: unknown SPI device"
-
-writeSpiByte :: forall w isa. (MachineWord w) => IoMem isa w -> Int -> Word8 -> Maybe (Either Text (IoMem isa w))
-writeSpiByte io addr byte =
-    findSpiByteRegister @w io addr <&> \(offset, wordAddr) ->
-        case offset of
-            0 -> fromMaybe (Left $ "iomemory[" <> show addr <> "]: unknown SPI device") $ writeSpiWord io wordAddr (byteToWord byte)
-            _ -> Left $ "iomemory[" <> show addr <> "]: can't write byte to SPI register part"
-
-popAvailableMiso :: Int -> [(w, Int)] -> Maybe ((w, Int), [(w, Int)])
-popAvailableMiso clock = go []
-    where
-        go _ [] = Nothing
-        go earlier (entry@(_, arrivalTick) : rest)
-            | arrivalTick <= clock = Just (entry, reverse earlier <> rest)
-            | otherwise = go (entry : earlier) rest
-
-spiPinsWord :: (Bits w, Num w) => Int -> IntMap (SpiDevice w) -> w
-spiPinsWord addr devices =
-    foldl' (.|.) 0 $ map (spiDevicePinsWord addr) (elems devices)
-
-spiDevicePinsWord :: (Bits w, Num w) => Int -> SpiDevice w -> w
-spiDevicePinsWord
-    addr
-    SpiDevice
-        { spiPins = SpiPinsConf{spCsAddr, spCsBit, spClkAddr, spClkBit, spMosiAddr, spMosiBit, spMisoAddr, spMisoBit}
-        , spiCsPin
-        , spiClkPin
-        , spiMosiPin
-        , spiMisoPin
-        } =
-        pinBitWord spCsAddr spCsBit spiCsPin
-            .|. pinBitWord spClkAddr spClkBit spiClkPin
-            .|. pinBitWord spMosiAddr spMosiBit spiMosiPin
-            .|. pinBitWord spMisoAddr spMisoBit spiMisoPin
-        where
-            pinBitWord pinAddr pinBit value =
-                if pinAddr == addr then bitWord pinBit value else 0
-
-bitWord :: (Bits w, Num w) => Int -> Bool -> w
-bitWord idx enabled = if enabled then bit idx else 0
-
-writeSpiPins :: forall w. (MachineWord w) => Int -> Int -> w -> SpiDevice w -> SpiDevice w
-writeSpiPins _clock addr word device =
-    let oldCs = spiCsPin device
-        oldClk = spiClkPin device
-        oldSoftClock = spiSoftClock device
-        SpiPinsConf{spCsAddr, spCsBit, spClkAddr, spClkBit, spMosiAddr, spMosiBit} = spiPins device
-        mode = spiClockMode device
-        newCs = updatePin spCsAddr spCsBit oldCs
-        newClk = updatePin spClkAddr spClkBit oldClk
-        newMosi = updatePin spMosiAddr spMosiBit (spiMosiPin device)
-        updatePin pinAddr pinBit oldValue =
-            if pinAddr == addr then testBit word pinBit else oldValue
-        deviceWithNewPins =
-            device
-                { spiCsPin = newCs
-                , spiClkPin = newClk
-                , spiMosiPin = newMosi
-                }
-        deviceAfterCs
-            | oldCs && not newCs =
-                if spiModePrimeOnActivate mode
-                    then
-                        primeMiso oldSoftClock deviceWithNewPins
-                    else
-                        deviceWithNewPins{spiMisoPin = False}
-            | newCs =
-                deviceWithNewPins
-                    { spiMisoPin = False
-                    , spiMosiShift = 0
-                    , spiMosiBits = 0
-                    }
-            | otherwise = deviceWithNewPins
-        active = not newCs
-        onRising = not oldClk && newClk
-        onFalling = oldClk && not newClk
-        sampleOnRising = spiModeSampleOnRising mode
-        sampleEdge = active && ((sampleOnRising && onRising) || (not sampleOnRising && onFalling))
-        shiftEdge = active && ((sampleOnRising && onFalling) || (not sampleOnRising && onRising))
-        afterShift =
-            if shiftEdge
-                then shiftMiso oldSoftClock deviceAfterCs
-                else deviceAfterCs
-     in if sampleEdge
-            then
-                let softClock' = oldSoftClock + 1
-                    sampled = shiftMosi softClock' newMosi afterShift
-                 in rememberSpiPins sampled{spiSoftClock = softClock'}
-            else rememberSpiPins afterShift
-
-rememberSpiPins :: SpiDevice w -> SpiDevice w
-rememberSpiPins device@SpiDevice{spiWaveLog} =
-    let snapshot = spiPinsSnapshot device
-     in case reverse spiWaveLog of
-            old : _ | samePins old snapshot -> device
-            _ -> device{spiWaveLog = spiWaveLog <> [snapshot]}
-
-spiPinsSnapshot :: SpiDevice w -> SpiPinsSnapshot
-spiPinsSnapshot SpiDevice{spiSoftClock, spiCsPin, spiClkPin, spiMosiPin, spiMisoPin} =
-    SpiPinsSnapshot
-        { spsTick = spiSoftClock
-        , spsCsPin = spiCsPin
-        , spsClkPin = spiClkPin
-        , spsMosiPin = spiMosiPin
-        , spsMisoPin = spiMisoPin
-        }
-
-samePins :: SpiPinsSnapshot -> SpiPinsSnapshot -> Bool
-samePins a b =
-    spsCsPin a
-        == spsCsPin b
-        && spsClkPin a
-        == spsClkPin b
-        && spsMosiPin a
-        == spsMosiPin b
-        && spsMisoPin a
-        == spsMisoPin b
-
-spiModeSampleOnRising :: SpiClockMode -> Bool
-spiModeSampleOnRising SpiMode0 = True
-spiModeSampleOnRising SpiMode1 = False
-spiModeSampleOnRising SpiMode2 = False
-spiModeSampleOnRising SpiMode3 = True
-
-spiModePrimeOnActivate :: SpiClockMode -> Bool
-spiModePrimeOnActivate SpiMode0 = True
-spiModePrimeOnActivate SpiMode1 = False
-spiModePrimeOnActivate SpiMode2 = True
-spiModePrimeOnActivate SpiMode3 = False
-
-primeMiso :: forall w. (MachineWord w) => Int -> SpiDevice w -> SpiDevice w
-primeMiso clock device =
-    let device' = ensureMisoLoaded clock device
-     in device'{spiMisoPin = currentMisoBit device'}
-
-ensureMisoLoaded :: forall w. (MachineWord w) => Int -> SpiDevice w -> SpiDevice w
-ensureMisoLoaded _clock device@SpiDevice{spiMisoShift}
-    | isJust spiMisoShift = device
-ensureMisoLoaded clock device@SpiDevice{spiMisoPending} =
-    case popAvailableMiso clock spiMisoPending of
-        Nothing -> device
-        Just ((w, tick), rest) ->
-            device
-                { spiMisoPending = rest
-                , spiMisoShift = Just SpiMisoShift{smsWord = w, smsBitIndex = finiteBitSize w - 1, smsTick = tick}
-                }
-
-currentMisoBit :: (Bits w) => SpiDevice w -> Bool
-currentMisoBit SpiDevice{spiMisoShift} =
-    case spiMisoShift of
-        Just SpiMisoShift{smsWord, smsBitIndex} -> testBit smsWord smsBitIndex
-        Nothing -> False
-
-shiftMiso :: forall w. (MachineWord w) => Int -> SpiDevice w -> SpiDevice w
-shiftMiso clock device =
-    let hadLoadedWord = isJust (spiMisoShift device)
-        device0 = ensureMisoLoaded clock device
-     in case spiMisoShift device0 of
-            Nothing -> device0{spiMisoPin = False}
-            _
-                | not hadLoadedWord ->
-                    device0{spiMisoPin = currentMisoBit device0}
-            Just SpiMisoShift{smsWord, smsBitIndex, smsTick} ->
-                if smsBitIndex == 0
-                    then
-                        let finished = (smsWord, smsTick)
-                            device1 =
-                                device0
-                                    { spiMisoShift = Nothing
-                                    , spiMisoConsumed = spiMisoConsumed device0 <> [finished]
-                                    }
-                            device2 = ensureMisoLoaded clock device1
-                         in device2{spiMisoPin = currentMisoBit device2}
-                    else
-                        let shift' = SpiMisoShift{smsWord, smsBitIndex = smsBitIndex - 1, smsTick}
-                            device1 = device0{spiMisoShift = Just shift'}
-                         in device1{spiMisoPin = currentMisoBit device1}
-
-shiftMosi :: forall w. (MachineWord w) => Int -> Bool -> SpiDevice w -> SpiDevice w
-shiftMosi clock mosiBit device@SpiDevice{spiMosiShift, spiMosiBits, spiMosiLog} =
-    let newBit :: w = if mosiBit then 1 else 0
-        nextWord = (spiMosiShift `shiftL` 1) .|. newBit
-        nextBits = spiMosiBits + 1
-        wordBits = finiteBitSize spiMosiShift
-     in if nextBits >= wordBits
-            then
-                device
-                    { spiMosiShift = 0
-                    , spiMosiBits = 0
-                    , spiMosiLog = spiMosiLog <> [(nextWord, clock)]
-                    }
-            else device{spiMosiShift = nextWord, spiMosiBits = nextBits}
 
 instance (ByteSize isa, MachineWord w, Memory (Mem isa w) isa w) => Memory (IoMem isa w) isa w where
     readInstruction io@IoMem{mIoStreams, mIoCells} idx =
