+++ test/golden/spi/spi_input_byte.risc-iv-32.result
@@ -1,7 +1,7 @@
 ---
 # Check SPI byte input
-spi_miso[0]: [] >>> [(165,0)]
-spi_mosi[0]: [] >>> [(165,32),(0,64)]
-spi_status[0]: miso_ready
-spi_clock[0]: 64
+spi_miso[0]: [] >>> [(165,0),(0,8)]
+spi_mosi[0]: [] >>> [(165,8),(0,16)]
+spi_status[0]: miso_empty
+spi_clock[0]: 16
 spi_pins[0]: cs=1 clk=0 mosi=0 miso=0
