+++ test/golden/spi/spi_input_bytes.risc-iv-32.result
@@ -1,7 +1,7 @@
 ---
 # Check SPI bytes input
-spi_miso[0]: [] >>> [(18,0),(52,1),(0,64)]
-spi_mosi[0]: [] >>> [(18,32),(52,64),(0,96)]
+spi_miso[0]: [] >>> [(18,0),(52,8),(0,16)]
+spi_mosi[0]: [] >>> [(18,8),(52,16),(0,24)]
 spi_status[0]: miso_empty
-spi_clock[0]: 96
+spi_clock[0]: 24
 spi_pins[0]: cs=1 clk=0 mosi=0 miso=0
