+++ test/golden/spi/spi_input_bytes.yaml
@@ -19,7 +19,7 @@ spi:
     input:
       - at: 0
         bytes: [0x12, 0x34]
-      - at: 64
+      - at: 16
         bytes: [0]
 reports:
   - name: Check SPI bytes input
@@ -33,8 +33,8 @@ reports:
       spi_clock[0]: {spi:0:clock}
       spi_pins[0]: {spi:0:pins}
     assert: |
-      spi_miso[0]: [] >>> [(18,0),(52,1),(0,64)]
-      spi_mosi[0]: [] >>> [(18,32),(52,64),(0,96)]
+      spi_miso[0]: [] >>> [(18,0),(52,8),(0,16)]
+      spi_mosi[0]: [] >>> [(18,8),(52,16),(0,24)]
       spi_status[0]: miso_empty
-      spi_clock[0]: 96
+      spi_clock[0]: 24
       spi_pins[0]: cs=1 clk=0 mosi=0 miso=0
